pyproject.toml
[tool.poetry]
name = "live-talk"
version = "0.1.0"
description = ""
authors = ["Digvijay Bhakuni <digvijay.bhakuni@tetruscorp.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.2"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
websockets = "^12.0"
openai-whisper = "^20231117"
rq = "^1.16.1"
transformers = "^4.40.1"
accelerate = "^0.29.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
